+++ Mainframe3270/__init__.py
@@ -74,6 +74,7 @@ class Mainframe3270(DynamicCore):
         wait_time_after_write: timedelta = timedelta(seconds=0),
         img_folder: str = ".",
         run_on_failure_keyword: str = "Take Screenshot",
+        model: str = "2",
     ) -> None:
         """
         By default the emulator visibility is set to visible=True.
@@ -96,7 +97,7 @@ class Mainframe3270(DynamicCore):
         self._running_on_failure_keyword = False
         self.register_run_on_failure_keyword(run_on_failure_keyword)
         libraries = [
-            X3270(visible, timeout, wait_time, wait_time_after_write, img_folder)
+            X3270(visible, timeout, wait_time, wait_time_after_write, img_folder, model)
         ]
         DynamicCore.__init__(self, libraries)
