+++ Mainframe3270/x3270.py
@@ -10,8 +10,7 @@ from typing import Any, List, Optional, Union
 from robot.api import logger
 from robot.api.deco import keyword
 from robot.libraries.BuiltIn import BuiltIn, RobotNotRunningError
-from robot.utils import (ConnectionCache, Matcher, secs_to_timestr,
-                         timestr_to_secs)
+from robot.utils import (ConnectionCache, secs_to_timestr,  timestr_to_secs)
 
 # fmt: on
 from .py3270 import Emulator
@@ -25,6 +24,7 @@ class X3270(object):
         wait_time: timedelta,
         wait_time_after_write: timedelta,
         img_folder: str,
+        model
     ) -> None:
         self.visible = visible
         self.timeout = self._convert_timeout(timeout)
@@ -32,6 +32,7 @@ class X3270(object):
         self.wait_write = self._convert_timeout(wait_time_after_write)
         self.imgfolder = img_folder
         self.cache = ConnectionCache()
+        self.model = model
         # Try Catch to run in Pycharm, and make a documentation in libdoc with no error
         try:
             self.output_folder = BuiltIn().get_variable_value("${OUTPUT DIR}")
@@ -105,7 +106,8 @@ class X3270(object):
             | Open Connection | Hostname | alias=my_first_connection |
         """
         extra_args = self._process_args(extra_args)
-        connection = Emulator(self.visible, self.timeout, extra_args)
+        model = self._get_model_from_list_or_file(extra_args)
+        connection = Emulator(self.visible, self.timeout, extra_args, model or self.model)
         host_string = f"{LU}@{host}" if LU else host
         if self._port_in_extra_args(extra_args):
             if port != 23:
@@ -134,6 +136,16 @@ class X3270(object):
                     for arg in shlex.split(line):
                         processed_args.append(arg)
         return processed_args
+
+    def _get_model_from_list_or_file(self, list_or_file):
+        pattern = re.compile(r"[wcxs3270.*]+model:\s*([327892345E-]+)")
+        match = None
+        if isinstance(list_or_file, list):
+            match = pattern.findall(str(list_or_file))
+        elif isinstance(list_or_file, os.PathLike) or isinstance(list_or_file, str):
+            with open(list_or_file) as file:
+                match = pattern.findall(file.read())
+        return None if not match else match[-1]
 
     def _port_in_extra_args(self, args) -> bool:
         if not args:
@@ -170,12 +182,12 @@ class X3270(object):
         """
         self._check_session_file_extension(session_file)
         self._check_contains_hostname(session_file)
-        self._check_model(session_file)
+        model = self._get_model_from_list_or_file(session_file)
         if os_name == "nt" and self.visible:
-            connection = Emulator(self.visible, self.timeout)
+            connection = Emulator(self.visible, self.timeout, model=model or self.model)
             connection.connect(str(session_file))
         else:
-            connection = Emulator(self.visible, self.timeout, [str(session_file)])
+            connection = Emulator(self.visible, self.timeout, [str(session_file)], model or self.model)
         return self.cache.register(connection, alias)
 
     def _check_session_file_extension(self, session_file):
@@ -202,22 +214,6 @@ class X3270(object):
                     "to set up the connection. "
                     "An example for wc3270 looks like this: \n"
                     "wc3270.hostname: myhost.com\n"
-                )
-
-    def _check_model(self, session_file):
-        with open(session_file) as file:
-            pattern = re.compile(r"[wcxs3270.*]+model:\s*([327892345E-]+)")
-            match = pattern.findall(file.read())
-            if not match:
-                return
-            elif match[-1] == "2":
-                return
-            else:
-                raise ValueError(
-                    f'Robot-Framework-Mainframe-3270-Library currently only supports model "2", '
-                    f'the model you specified in your session file was "{match[-1]}". '
-                    f'Please change it to "2", using either the session wizard if you are on Windows, '
-                    f'or by editing the model resource like this "*model: 2"'
                 )
 
     @keyword("Switch Connection")
@@ -287,41 +283,6 @@ class X3270(object):
             | Change Wait Time After Write | 0:00:02       |
         """
         self.wait_write = self._convert_timeout(wait_time_after_write)
-
-    @keyword("Read")
-    def read(self, ypos: int, xpos: int, length: int) -> str:
-        """Get a string of ``length`` at screen co-ordinates ``ypos`` / ``xpos``.
-
-        Co-ordinates are 1 based, as listed in the status area of the terminal.
-
-        Example for read a string in the position y=8 / x=10 of a length 15:
-            | ${value} | Read | 8 | 10 | 15 |
-        """
-        self._check_limits(ypos, xpos)
-        # Checks if the user has passed a length that will be larger than the x limit of the screen.
-        if (xpos + length) > (80 + 1):
-            raise Exception(
-                "You have exceeded the x-axis limit of the mainframe screen"
-            )
-        string = self.mf.string_get(ypos, xpos, length)
-        return string
-
-    @keyword("Read All Screen")
-    def read_all_screen(self) -> str:
-        """Read the current screen and returns all content in one string.
-
-        This is useful if your automation scripts should take different routes depending
-        on a message shown on the screen.
-
-        Example:
-            | ${screen} | Read All Screen              |
-            | IF   | 'certain text' in '''${screen}''' |
-            |      | Do Something                      |
-            | ELSE |                                   |
-            |      | Do Something Else                 |
-            | END  |                                   |
-        """
-        return self._read_all_screen()
 
     @keyword("Execute Command")
     def execute_command(self, cmd: str) -> None:
@@ -515,10 +476,10 @@ class X3270(object):
     ) -> None:
         txt = txt.encode("unicode_escape")
         if ypos is not None and xpos is not None:
-            self._check_limits(ypos, xpos)
-            self.mf.send_string(txt, ypos, xpos)
+            Emulator()._check_limits(ypos, xpos)
+            Emulator().send_string(txt, ypos, xpos)
         else:
-            self.mf.send_string(txt)
+            Emulator().send_string(txt)
         time.sleep(self.wait_write)
         for i in range(enter):
             self.mf.send_enter()
@@ -540,344 +501,8 @@ class X3270(object):
         timeout = self._convert_timeout(timeout)
         max_time = time.time() + timeout  # type: ignore
         while time.time() < max_time:
-            result = self._search_string(str(txt))
+            result = Emulator()._search_string(str(txt))
             if result:
                 return txt
         raise Exception(f'String "{txt}" not found in {secs_to_timestr(timeout)}')
 
-    def _search_string(self, string: str, ignore_case: bool = False) -> bool:
-        """Search if a string exists on the mainframe screen and return True or False."""
-
-        def __read_screen(string: str, ignore_case: bool) -> bool:
-            for ypos in range(24):
-                line = self.mf.string_get(ypos + 1, 1, 80)
-                if ignore_case:
-                    line = line.lower()
-                if string in line:
-                    return True
-            return False
-
-        status = __read_screen(string, ignore_case)
-        return status
-
-    @keyword("Page Should Contain String")
-    def page_should_contain_string(
-        self, txt: str, ignore_case: bool = False, error_message: Optional[str] = None
-    ) -> None:
-        """Assert that a given string exists on the mainframe screen.
-
-        The assertion is case sensitive. If you want it to be case insensitive, you can pass the argument ignore_case=True.
-
-        You can change the exception message by setting a custom string to error_message.
-
-        Example:
-            | Page Should Contain String | something |
-            | Page Should Contain String | someTHING | ignore_case=True                |
-            | Page Should Contain String | something | error_message=New error message |
-        """
-        message = f'The string "{txt}" was not found'
-        if error_message:
-            message = error_message
-        if ignore_case:
-            txt = txt.lower()
-        result = self._search_string(txt, ignore_case)
-        if not result:
-            raise Exception(message)
-        logger.info(f'The string "{txt}" was found')
-
-    @keyword("Page Should Not Contain String")
-    def page_should_not_contain_string(
-        self, txt: str, ignore_case: bool = False, error_message: Optional[str] = None
-    ) -> None:
-        """Assert that a given string does NOT exists on the mainframe screen.
-
-        The assertion is case sensitive. If you want it to be case insensitive, you can pass the argument ignore_case=True.
-
-        You can change the exception message by setting a custom string to error_message.
-
-        Example:
-            | Page Should Not Contain String | something |
-            | Page Should Not Contain String | someTHING | ignore_case=True |
-            | Page Should Not Contain String | something | error_message=New error message |
-        """
-        message = f'The string "{txt}" was found'
-        if error_message:
-            message = error_message
-        if ignore_case:
-            txt = txt.lower()
-        result = self._search_string(txt, ignore_case)
-        if result:
-            raise Exception(message)
-
-    @keyword("Page Should Contain Any String")
-    def page_should_contain_any_string(
-        self,
-        list_string: List[str],
-        ignore_case: bool = False,
-        error_message: Optional[str] = None,
-    ) -> None:
-        """Assert that one of the strings in a given list exists on the mainframe screen.
-
-        The assertion is case sensitive. If you want it to be case insensitive, you can pass the argument ignore_case=True.
-
-        You can change the exception message by setting a custom string to error_message.
-
-        Example:
-            | Page Should Contain Any String | ${list_of_string} |
-            | Page Should Contain Any String | ${list_of_string} | ignore_case=True |
-            | Page Should Contain Any String | ${list_of_string} | error_message=New error message |
-        """
-        message = f'The strings "{list_string}" were not found'
-        if error_message:
-            message = error_message
-        if ignore_case:
-            list_string = [item.lower() for item in list_string]
-        for string in list_string:
-            result = self._search_string(string, ignore_case)
-            if result:
-                break
-        if not result:
-            raise Exception(message)
-
-    @keyword("Page Should Not Contain Any String")
-    def page_should_not_contain_any_string(
-        self,
-        list_string: List[str],
-        ignore_case: bool = False,
-        error_message: Optional[str] = None,
-    ) -> None:
-        """Assert that none of the strings in a given list exists on the mainframe screen. If one or more of the
-        string are found, the keyword will raise a exception.
-
-        The assertion is case sensitive. If you want it to be case insensitive, you can pass the argument ignore_case=True.
-
-        You can change the exception message by setting a custom string to error_message.
-
-        Example:
-            | Page Should Not Contain Any Strings | ${list_of_string} |
-            | Page Should Not Contain Any Strings | ${list_of_string} | ignore_case=True |
-            | Page Should Not Contain Any Strings | ${list_of_string} | error_message=New error message |
-        """
-        self._compare_all_list_with_screen_text(
-            list_string, ignore_case, error_message, should_match=False
-        )
-
-    @keyword("Page Should Contain All Strings")
-    def page_should_contain_all_strings(
-        self,
-        list_string: List[str],
-        ignore_case: bool = False,
-        error_message: Optional[str] = None,
-    ) -> None:
-        """Assert that all of the strings in a given list exist on the mainframe screen.
-
-        The assertion is case sensitive. If you want it to be case insensitive, you can pass the argument ignore_case=True.
-
-        You can change the exception message by setting a custom string to error_message.
-
-        Example:
-            | Page Should Contain All Strings | ${list_of_string} |
-            | Page Should Contain All Strings | ${list_of_string} | ignore_case=True |
-            | Page Should Contain All Strings | ${list_of_string} | error_message=New error message |
-        """
-        self._compare_all_list_with_screen_text(
-            list_string, ignore_case, error_message, should_match=True
-        )
-
-    @keyword("Page Should Not Contain All Strings")
-    def page_should_not_contain_all_strings(
-        self,
-        list_string: List[str],
-        ignore_case: bool = False,
-        error_message: Optional[str] = None,
-    ) -> None:
-        """Fails if one of the strings in a given list exists on the mainframe screen. If one of the string
-        are found, the keyword will raise a exception.
-
-        The assertion is case sensitive. If you want it to be case insensitive, you can pass the argument ignore_case=True.
-
-        You can change the exception message by setting a custom string to error_message.
-
-        Example:
-            | Page Should Not Contain All Strings | ${list_of_string} |
-            | Page Should Not Contain All Strings | ${list_of_string} | ignore_case=True |
-            | Page Should Not Contain All Strings | ${list_of_string} | error_message=New error message |
-        """
-        message = error_message
-        if ignore_case:
-            list_string = [item.lower() for item in list_string]
-        for string in list_string:
-            result = self._search_string(string, ignore_case)
-            if result:
-                if message is None:
-                    message = f'The string "{string}" was found'
-                raise Exception(message)
-
-    @keyword("Page Should Contain String X Times")
-    def page_should_contain_string_x_times(
-        self,
-        txt: str,
-        number: int,
-        ignore_case: bool = False,
-        error_message: Optional[str] = None,
-    ) -> None:
-        """Asserts that the entered string appears the desired number of times on the mainframe screen.
-
-        The assertion is case sensitive. If you want it to be case insensitive, you can pass the argument ignore_case=True.
-
-        You can change the exception message by setting a custom string to error_message.
-
-        Example:
-               | Page Should Contain String X Times | something | 3 |
-               | Page Should Contain String X Times | someTHING | 3 | ignore_case=True |
-               | Page Should Contain String X Times | something | 3 | error_message=New error message |
-        """
-        message = error_message
-        number = number
-        all_screen = self._read_all_screen()
-        if ignore_case:
-            txt = txt.lower()
-            all_screen = all_screen.lower()
-        number_of_times = all_screen.count(txt)
-        if number_of_times != number:
-            if message is None:
-                message = f'The string "{txt}" was not found "{number}" times, it appears "{number_of_times}" times'
-            raise Exception(message)
-        logger.info(f'The string "{txt}" was found "{number}" times')
-
-    @keyword("Page Should Match Regex")
-    def page_should_match_regex(self, regex_pattern: str) -> None:
-        r"""Fails if string does not match pattern as a regular expression. Regular expression check is
-        implemented using the Python [https://docs.python.org/2/library/re.html|re module]. Python's
-        regular expression syntax is derived from Perl, and it is thus also very similar to the syntax used,
-        for example, in Java, Ruby and .NET.
-
-        Backslash is an escape character in the test data, and possible backslashes in the pattern must
-        thus be escaped with another backslash (e.g. \\d\\w+).
-        """
-        page_text = self._read_all_screen()
-        if not re.findall(regex_pattern, page_text, re.MULTILINE):
-            raise Exception(f'No matches found for "{regex_pattern}" pattern')
-
-    @keyword("Page Should Not Match Regex")
-    def page_should_not_match_regex(self, regex_pattern: str) -> None:
-        r"""Fails if string does match pattern as a regular expression. Regular expression check is
-        implemented using the Python [https://docs.python.org/2/library/re.html|re module]. Python's
-        regular expression syntax is derived from Perl, and it is thus also very similar to the syntax used,
-        for example, in Java, Ruby and .NET.
-
-        Backslash is an escape character in the test data, and possible backslashes in the pattern must
-        thus be escaped with another backslash (e.g. \\d\\w+).
-        """
-        page_text = self._read_all_screen()
-        if re.findall(regex_pattern, page_text, re.MULTILINE):
-            raise Exception(f'There are matches found for "{regex_pattern}" pattern')
-
-    @keyword("Page Should Contain Match")
-    def page_should_contain_match(
-        self, txt: str, ignore_case: bool = False, error_message: Optional[str] = None
-    ) -> None:
-        """Assert that the text displayed on the mainframe screen matches the given pattern.
-
-        Pattern matching is similar to matching files in a shell, and it is always case sensitive.
-        In the pattern, * matches anything and ? matches any single character.
-
-        Note that for this keyword the entire screen is considered a string. So if you want to search
-        for the string "something" and it is somewhere other than at the beginning or end of the screen, it
-        should be reported as follows: **something**
-
-        The assertion is case sensitive. If you want it to be case insensitive, you can pass the argument ignore_case=True.
-
-        You can change the exception message by setting a custom string to error_message.
-
-        Example:
-            | Page Should Contain Match | **something** |
-            | Page Should Contain Match | **so???hing** |
-            | Page Should Contain Match | **someTHING** | ignore_case=True |
-            | Page Should Contain Match | **something** | error_message=New error message |
-        """
-        message = error_message
-        all_screen = self._read_all_screen()
-        if ignore_case:
-            txt = txt.lower()
-            all_screen = all_screen.lower()
-        matcher = Matcher(txt, caseless=False, spaceless=False)
-        result = matcher.match(all_screen)
-        if not result:
-            if message is None:
-                message = f'No matches found for "{txt}" pattern'
-            raise Exception(message)
-
-    @keyword("Page Should Not Contain Match")
-    def page_should_not_contain_match(
-        self, txt: str, ignore_case: bool = False, error_message: Optional[str] = None
-    ) -> None:
-        """Assert that the text displayed on the mainframe screen does NOT match the given pattern.
-
-        Pattern matching is similar to matching files in a shell, and it is always case sensitive.
-        In the pattern, * matches anything and ? matches any single character.
-
-        Note that for this keyword the entire screen is considered a string. So if you want to search
-        for the string "something" and it is somewhere other than at the beginning or end of the screen, it
-        should be reported as follows: **something**
-
-        The assertion is case sensitive. If you want it to be case insensitive, you can pass the argument ignore_case=True.
-
-        You can change the exception message by setting a custom string to error_message.
-
-        Example:
-            | Page Should Not Contain Match | **something** |
-            | Page Should Not Contain Match | **so???hing** |
-            | Page Should Not Contain Match | **someTHING** | ignore_case=True |
-            | Page Should Not Contain Match | **something** | error_message=New error message |
-        """
-        message = error_message
-        all_screen = self._read_all_screen()
-        if ignore_case:
-            txt = txt.lower()
-            all_screen = all_screen.lower()
-        matcher = Matcher(txt, caseless=False, spaceless=False)
-        result = matcher.match(all_screen)
-        if result:
-            if message is None:
-                message = f'There are matches found for "{txt}" pattern'
-            raise Exception(message)
-
-    def _read_all_screen(self) -> str:
-        """Read all the mainframe screen and return in a single string."""
-        full_text = ""
-        for ypos in range(24):
-            full_text += self.mf.string_get(ypos + 1, 1, 80)
-        return full_text
-
-    def _compare_all_list_with_screen_text(
-        self,
-        list_string: List[str],
-        ignore_case: bool,
-        message: Optional[str],
-        should_match: bool,
-    ) -> None:
-        if ignore_case:
-            list_string = [item.lower() for item in list_string]
-        for string in list_string:
-            result = self._search_string(string, ignore_case)
-            if not should_match and result:
-                if message is None:
-                    message = f'The string "{string}" was found'
-                raise Exception(message)
-            elif should_match and not result:
-                if message is None:
-                    message = f'The string "{string}" was not found'
-                raise Exception(message)
-
-    @staticmethod
-    def _check_limits(ypos: int, xpos: int):
-        """Checks if the user has passed some coordinate y / x greater than that existing in the mainframe"""
-        if ypos > 24:
-            raise Exception(
-                "You have exceeded the y-axis limit of the mainframe screen"
-            )
-        if xpos > 80:
-            raise Exception(
-                "You have exceeded the x-axis limit of the mainframe screen"
-            )
